Suggestions from Nestor
